added: 832, 894A, 908A, 1343A, 1369A, 1385A, 1400A

+++ A/832A.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    int n, k;
+    long long n, k;
     cin >> n >> k;
 
     cout << ((n / k) % 2 == 0 ? "NO" : "YES");
